pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Sesiunina Olga <sesyunina.olga7@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
gunicorn = "^22.0.0"
urllib3 = "1.26.15"
psycopg = "^3.1.18"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
validators = "^0.28.1"
bs4 = "^0.0.2"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
